1.1.0: Subject state methods beta/remove
