+++ src/main/java/rx/subjects/AsyncSubject.java
@@ -15,11 +15,10 @@
  */
 package rx.subjects;
 
-import java.lang.reflect.Array;
 import java.util.*;
 
 import rx.Observer;
-import rx.annotations.Experimental;
+import rx.annotations.Beta;
 import rx.exceptions.Exceptions;
 import rx.functions.Action1;
 import rx.internal.operators.NotificationLite;
@@ -141,8 +140,7 @@ public final class AsyncSubject<T> extends Subject<T, T> {
      * retrieved by {@code getValue()} may get outdated.
      * @return true if and only if the subject has some value but not an error
      */
-    @Experimental
-    @Override
+    @Beta
     public boolean hasValue() {
         Object v = lastValue;
         Object o = state.get();
@@ -152,8 +150,7 @@ public final class AsyncSubject<T> extends Subject<T, T> {
      * Check if the Subject has terminated with an exception.
      * @return true if the subject has received a throwable through {@code onError}.
      */
-    @Experimental
-    @Override
+    @Beta
     public boolean hasThrowable() {
         Object o = state.get();
         return nl.isError(o);
@@ -162,8 +159,7 @@ public final class AsyncSubject<T> extends Subject<T, T> {
      * Check if the Subject has terminated normally.
      * @return true if the subject completed normally via {@code onCompleted()}
      */
-    @Experimental
-    @Override
+    @Beta
     public boolean hasCompleted() {
         Object o = state.get();
         return o != null && !nl.isError(o);
@@ -177,8 +173,7 @@ public final class AsyncSubject<T> extends Subject<T, T> {
      * @return the current value or {@code null} if the Subject doesn't have a value,
      * has terminated with an exception or has an actual {@code null} as a value.
      */
-    @Experimental
-    @Override
+    @Beta
     public T getValue() {
         Object v = lastValue;
         Object o = state.get();
@@ -192,34 +187,12 @@ public final class AsyncSubject<T> extends Subject<T, T> {
      * @return the Throwable that terminated the Subject or {@code null} if the
      * subject hasn't terminated yet or it terminated normally.
      */
-    @Experimental
-    @Override
+    @Beta
     public Throwable getThrowable() {
         Object o = state.get();
         if (nl.isError(o)) {
             return nl.getError(o);
         }
         return null;
-    }
-    @Override
-    @Experimental
-    @SuppressWarnings("unchecked")
-    public T[] getValues(T[] a) {
-        Object v = lastValue;
-        Object o = state.get();
-        if (!nl.isError(o) && nl.isNext(v)) {
-            T val = nl.getValue(v);
-            if (a.length == 0) {
-                a = (T[])Array.newInstance(a.getClass().getComponentType(), 1);
-            }
-            a[0] = val;
-            if (a.length > 1) {
-                a[1] = null;
-            }
-        } else
-        if (a.length > 0) {
-            a[0] = null;
-        }
-        return a;
     }
 }
