+++ src/main/java/rx/subjects/BehaviorSubject.java
@@ -20,7 +20,7 @@ import java.lang.reflect.Array;
 import java.util.*;
 
 import rx.Observer;
-import rx.annotations.Experimental;
+import rx.annotations.Beta;
 import rx.exceptions.Exceptions;
 import rx.functions.Action1;
 import rx.internal.operators.NotificationLite;
@@ -177,8 +177,7 @@ public final class BehaviorSubject<T> extends Subject<T, T> {
      * retrieved by {@code getValue()} may get outdated.
      * @return true if and only if the subject has some value and hasn't terminated yet.
      */
-    @Experimental
-    @Override
+    @Beta
     public boolean hasValue() {
         Object o = state.get();
         return nl.isNext(o);
@@ -187,8 +186,7 @@ public final class BehaviorSubject<T> extends Subject<T, T> {
      * Check if the Subject has terminated with an exception.
      * @return true if the subject has received a throwable through {@code onError}.
      */
-    @Experimental
-    @Override
+    @Beta
     public boolean hasThrowable() {
         Object o = state.get();
         return nl.isError(o);
@@ -197,8 +195,7 @@ public final class BehaviorSubject<T> extends Subject<T, T> {
      * Check if the Subject has terminated normally.
      * @return true if the subject completed normally via {@code onCompleted()}
      */
-    @Experimental
-    @Override
+    @Beta
     public boolean hasCompleted() {
         Object o = state.get();
         return nl.isCompleted(o);
@@ -212,8 +209,7 @@ public final class BehaviorSubject<T> extends Subject<T, T> {
      * @return the current value or {@code null} if the Subject doesn't have a value,
      * has terminated or has an actual {@code null} as a valid value.
      */
-    @Experimental
-    @Override
+    @Beta
     public T getValue() {
         Object o = state.get();
         if (nl.isNext(o)) {
@@ -226,8 +222,7 @@ public final class BehaviorSubject<T> extends Subject<T, T> {
      * @return the Throwable that terminated the Subject or {@code null} if the
      * subject hasn't terminated yet or it terminated normally.
      */
-    @Experimental
-    @Override
+    @Beta
     public Throwable getThrowable() {
         Object o = state.get();
         if (nl.isError(o)) {
@@ -235,8 +230,13 @@ public final class BehaviorSubject<T> extends Subject<T, T> {
         }
         return null;
     }
-    @Override
-    @Experimental
+    /**
+     * Returns a snapshot of the currently buffered non-terminal events into 
+     * the provided {@code a} array or creates a new array if it has not enough capacity.
+     * @param a the array to fill in
+     * @return the array {@code a} if it had enough capacity or a new array containing the available values 
+     */
+    @Beta
     @SuppressWarnings("unchecked")
     public T[] getValues(T[] a) {
         Object o = state.get();
@@ -253,5 +253,25 @@ public final class BehaviorSubject<T> extends Subject<T, T> {
             a[0] = null;
         }
         return a;
+    }
+    
+    /** An empty array to trigger getValues() to return a new array. */
+    private static final Object[] EMPTY_ARRAY = new Object[0];
+    
+    /**
+     * Returns a snapshot of the currently buffered non-terminal events.
+     * <p>The operation is threadsafe.
+     *
+     * @return a snapshot of the currently buffered non-terminal events.
+     * @since (If this graduates from being an Experimental class method, replace this parenthetical with the release number)
+     */
+    @SuppressWarnings("unchecked")
+    @Beta
+    public Object[] getValues() {
+        T[] r = getValues((T[])EMPTY_ARRAY);
+        if (r == EMPTY_ARRAY) {
+            return new Object[0]; // don't leak the default empty array.
+        }
+        return r;
     }
 }
