+++ src/main/java/rx/subjects/PublishSubject.java
@@ -18,7 +18,7 @@ package rx.subjects;
 import java.util.*;
 
 import rx.Observer;
-import rx.annotations.Experimental;
+import rx.annotations.Beta;
 import rx.exceptions.Exceptions;
 import rx.functions.Action1;
 import rx.internal.operators.NotificationLite;
@@ -124,8 +124,7 @@ public final class PublishSubject<T> extends Subject<T, T> {
      * Check if the Subject has terminated with an exception.
      * @return true if the subject has received a throwable through {@code onError}.
      */
-    @Experimental
-    @Override
+    @Beta
     public boolean hasThrowable() {
         Object o = state.get();
         return nl.isError(o);
@@ -134,8 +133,7 @@ public final class PublishSubject<T> extends Subject<T, T> {
      * Check if the Subject has terminated normally.
      * @return true if the subject completed normally via {@code onCompleted}
      */
-    @Experimental
-    @Override
+    @Beta
     public boolean hasCompleted() {
         Object o = state.get();
         return o != null && !nl.isError(o);
@@ -145,37 +143,12 @@ public final class PublishSubject<T> extends Subject<T, T> {
      * @return the Throwable that terminated the Subject or {@code null} if the
      * subject hasn't terminated yet or it terminated normally.
      */
-    @Experimental
-    @Override
+    @Beta
     public Throwable getThrowable() {
         Object o = state.get();
         if (nl.isError(o)) {
             return nl.getError(o);
         }
         return null;
-    }
-    
-    @Override
-    @Experimental
-    public boolean hasValue() {
-        return false;
-    }
-    @Override
-    @Experimental
-    public T getValue() {
-        return null;
-    }
-    @Override
-    @Experimental
-    public Object[] getValues() {
-        return new Object[0];
-    }
-    @Override
-    @Experimental
-    public T[] getValues(T[] a) {
-        if (a.length > 0) {
-            a[0] = null;
-        }
-        return a;
     }
 }
