+++ src/main/java/rx/subjects/ReplaySubject.java
@@ -22,7 +22,7 @@ import java.util.concurrent.atomic.AtomicIntegerFieldUpdater;
 
 import rx.*;
 import rx.Observer;
-import rx.annotations.Experimental;
+import rx.annotations.Beta;
 import rx.exceptions.Exceptions;
 import rx.functions.*;
 import rx.internal.operators.NotificationLite;
@@ -1098,8 +1098,7 @@ public final class ReplaySubject<T> extends Subject<T, T> {
      * Check if the Subject has terminated with an exception.
      * @return true if the subject has received a throwable through {@code onError}.
      */
-    @Experimental
-    @Override
+    @Beta
     public boolean hasThrowable() {
         NotificationLite<T> nl = ssm.nl;
         Object o = ssm.get();
@@ -1109,8 +1108,7 @@ public final class ReplaySubject<T> extends Subject<T, T> {
      * Check if the Subject has terminated normally.
      * @return true if the subject completed normally via {@code onCompleted}
      */
-    @Experimental
-    @Override
+    @Beta
     public boolean hasCompleted() {
         NotificationLite<T> nl = ssm.nl;
         Object o = ssm.get();
@@ -1121,8 +1119,7 @@ public final class ReplaySubject<T> extends Subject<T, T> {
      * @return the Throwable that terminated the Subject or {@code null} if the
      * subject hasn't terminated yet or it terminated normally.
      */
-    @Experimental
-    @Override
+    @Beta
     public Throwable getThrowable() {
         NotificationLite<T> nl = ssm.nl;
         Object o = ssm.get();
@@ -1135,19 +1132,18 @@ public final class ReplaySubject<T> extends Subject<T, T> {
      * Returns the current number of items (non-terminal events) available for replay.
      * @return the number of items available
      */
-    @Experimental
+    @Beta
     public int size() {
         return state.size();
     }
     /**
      * @return true if the Subject holds at least one non-terminal event available for replay
      */
-    @Experimental
+    @Beta
     public boolean hasAnyValue() {
         return !state.isEmpty();
     }
-    @Experimental
-    @Override
+    @Beta
     public boolean hasValue() {
         return hasAnyValue();
     }
@@ -1157,12 +1153,32 @@ public final class ReplaySubject<T> extends Subject<T, T> {
      * @param a the array to fill in
      * @return the array {@code a} if it had enough capacity or a new array containing the available values 
      */
-    @Experimental
-    @Override
+    @Beta
     public T[] getValues(T[] a) {
         return state.toArray(a);
     }
-    @Override
+    
+    /** An empty array to trigger getValues() to return a new array. */
+    private static final Object[] EMPTY_ARRAY = new Object[0];
+    
+    /**
+     * Returns a snapshot of the currently buffered non-terminal events.
+     * <p>The operation is threadsafe.
+     *
+     * @return a snapshot of the currently buffered non-terminal events.
+     * @since (If this graduates from being an Experimental class method, replace this parenthetical with the release number)
+     */
+    @SuppressWarnings("unchecked")
+    @Beta
+    public Object[] getValues() {
+        T[] r = getValues((T[])EMPTY_ARRAY);
+        if (r == EMPTY_ARRAY) {
+            return new Object[0]; // don't leak the default empty array.
+        }
+        return r;
+    }
+    
+    @Beta
     public T getValue() {
         return state.latest();
     }
